Rename ConfirmedCasesTable to ConfirmedCasesDetailsTable
