+++ components/ConfirmedCasesDetailsTable.vue
